[refactor] 상품 리스트 조회 응답 객체 서비스로 이동
