+++ src/main/java/kr/hhplus/be/server/product/controller/ProductController.java
@@ -2,6 +2,7 @@ package kr.hhplus.be.server.product.controller;
 
 import kr.hhplus.be.server.ApiResponse;
 import kr.hhplus.be.server.product.service.ProductReadService;
+import kr.hhplus.be.server.product.service.dto.ProductResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,7 +17,7 @@ import java.util.List;
 public class ProductController {
     private final ProductReadService productReadService;
 
-    @GetMapping()
+    @GetMapping
     public ApiResponse<List<ProductResponse>> getProducts(@RequestParam int page, @RequestParam int size) {
         return ApiResponse.ok(productReadService.getProducts(page, size));
     }
