+++ src/main/java/kr/hhplus/be/server/product/service/ProductReadService.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.product.service;
 
-import kr.hhplus.be.server.product.controller.ProductResponse;
+import kr.hhplus.be.server.product.service.dto.ProductResponse;
 import kr.hhplus.be.server.product.domain.Product;
 import kr.hhplus.be.server.product.domain.ProductStock;
 import kr.hhplus.be.server.product.repository.ProductRepository;
